// src/models/PlaceAutoComplete.ts
// Generated by https://quicktype.io

export default interface PlaceAutoComplete {
  predictions: Prediction[];
  status: string;
}

export interface Prediction {
  description: string;
  matched_substrings: MatchedSubstring[];
  place_id: string;
  reference: string;
  structured_formatting: StructuredFormatting;
  terms: Term[];
  types: Type[];
}

interface MatchedSubstring {
  length: number;
  offset: number;
}

interface StructuredFormatting {
  main_text: string;
  main_text_matched_substrings: MatchedSubstring[];
  secondary_text: string;
}

interface Term {
  offset: number;
  value: string;
}

enum Type {
  Geocode = 'geocode',
  Route = 'route',
}